# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Doe Me API",
        "description": "API para o aplicativo Doe Me - plataforma de doações que conecta pessoas dispostas a doar bens usados com pessoas que precisam deles",
        "contact": {
            "name": "Equipe Doe Me",
            "email": "contato@doeme.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor de desenvolvimento"
        },
        {
            "url": "https://api.doeme.com",
            "description": "Servidor de produção"
        }
    ],
    "paths": {
...
        "/api/donation-items": {
            "get": {
                "tags": [
                    "Donation Items"
                ],
                "summary": "Listar itens de doação",
                "operationId": "2982508ca5e3f22a098a0d6ba0c4b7e7",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "available",
                                "reserved",
                                "donated"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de itens de doação"
                    }
                }
            },
            "post": {
                "tags": [
                    "Donation Items"
                ],
                "summary": "Criar novo item de doação",
                "operationId": "d2c876437a10beadf7515917a2e0327b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "category_id",
                                    "condition",
                                    "location"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "condition": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "latitude": {
                                        "type": "number"
                                    },
                                    "longitude": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item criado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/donation-items/{id}": {
            "get": {
                "tags": [
                    "Donation Items"
                ],
                "summary": "Obter item específico",
                "operationId": "7184c0a02889cd14167d68d4476b6236",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do item"
                    }
                }
            },
            "put": {
                "tags": [
                    "Donation Items"
                ],
                "summary": "Atualizar item de doação",
                "operationId": "62ece3b4d3667340e413338ba13f04a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item atualizado com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Donation Items"
                ],
                "summary": "Excluir item de doação",
                "operationId": "92f51b5f32f20d7f04dd9e917c3f5a33",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item excluído com sucesso"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-donations": {
            "get": {
                "tags": [
                    "Donation Items"
                ],
                "summary": "Listar meus itens de doação",
                "operationId": "75103b38f2f92e382b7db24d88af9afb",
                "responses": {
                    "200": {
                        "description": "Lista dos meus itens"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Bearer",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}